chore(insights): move stickiness charts out of trends folder
Stickiness is its own InsightType, so its chart components no longer
live under `products/product_analytics/frontend/insights/trends/`.
They now sit alongside `trends/`, `funnels/`, and `retention/` at
`insights/stickiness/`. Imports of the still-shared `chartErrorHandler`,
`trendsSeriesMeta`, and `TrendsTooltip` (consumed by trends, stickiness,
and lifecycle) are repointed; nothing else changes.

Generated-By: PostHog Code
Task-Id: 2a9cf3d4-124d-4d01-a3db-0bc68b0fc81f
