+++ frontend/src/scenes/trends/Trends.tsx
@@ -35,14 +35,18 @@ const TrendsBarChart = lazy(() =>
     }))
 )
 const StickinessLineChart = lazy(() =>
-    import('products/product_analytics/frontend/insights/trends/StickinessLineChart/StickinessLineChart').then((m) => ({
-        default: m.StickinessLineChart,
-    }))
+    import('products/product_analytics/frontend/insights/stickiness/StickinessLineChart/StickinessLineChart').then(
+        (m) => ({
+            default: m.StickinessLineChart,
+        })
+    )
 )
 const StickinessBarChart = lazy(() =>
-    import('products/product_analytics/frontend/insights/trends/StickinessBarChart/StickinessBarChart').then((m) => ({
-        default: m.StickinessBarChart,
-    }))
+    import('products/product_analytics/frontend/insights/stickiness/StickinessBarChart/StickinessBarChart').then(
+        (m) => ({
+            default: m.StickinessBarChart,
+        })
+    )
 )
 const TrendsPieChart = lazy(() =>
     import('products/product_analytics/frontend/insights/trends/TrendsPieChart/TrendsPieChart').then((m) => ({
