+++ products/product_analytics/frontend/insights/stickiness/StickinessBarChart/StickinessBarChart.test.tsx
