+++ products/product_analytics/frontend/insights/stickiness/StickinessBarChart/stickinessBarChartTransforms.ts
