+++ products/product_analytics/frontend/insights/stickiness/StickinessLineChart/StickinessLineChart.test.tsx
