+++ products/product_analytics/frontend/insights/stickiness/StickinessLineChart/handleStickinessChartClick.tsx
