+++ products/product_analytics/frontend/insights/stickiness/StickinessLineChart/stickinessChartTransforms.ts
